Make mockbroker's ID an int32 to avoid a boatload of casts

+++ client_test.go
@@ -31,7 +31,7 @@ func TestClientExtraBrokers(t *testing.T) {
 	mb2 := NewMockBroker(t, 2)
 
 	mdr := new(MetadataResponse)
-	mdr.AddBroker(mb2.Addr(), int32(mb2.BrokerID()))
+	mdr.AddBroker(mb2.Addr(), mb2.BrokerID())
 	mb1.Returns(mdr)
 
 	client, err := NewClient("client_id", []string{mb1.Addr()}, nil)
@@ -49,8 +49,8 @@ func TestClientMetadata(t *testing.T) {
 	mb5 := NewMockBroker(t, 5)
 
 	mdr := new(MetadataResponse)
-	mdr.AddBroker(mb5.Addr(), int32(mb5.BrokerID()))
-	mdr.AddTopicPartition("my_topic", 0, int32(mb5.BrokerID()))
+	mdr.AddBroker(mb5.Addr(), mb5.BrokerID())
+	mdr.AddTopicPartition("my_topic", 0, mb5.BrokerID())
 	mb1.Returns(mdr)
 
 	client, err := NewClient("client_id", []string{mb1.Addr()}, nil)
@@ -88,11 +88,11 @@ func TestClientRefreshBehaviour(t *testing.T) {
 	mb5 := NewMockBroker(t, 5)
 
 	mdr := new(MetadataResponse)
-	mdr.AddBroker(mb5.Addr(), int32(mb5.BrokerID()))
+	mdr.AddBroker(mb5.Addr(), mb5.BrokerID())
 	mb1.Returns(mdr)
 
 	mdr2 := new(MetadataResponse)
-	mdr2.AddTopicPartition("my_topic", 0xb, int32(mb5.BrokerID()))
+	mdr2.AddTopicPartition("my_topic", 0xb, mb5.BrokerID())
 	mb5.Returns(mdr2)
 
 	client, err := NewClient("clientID", []string{mb1.Addr()}, nil)
